backend/requirements.txt
fastapi[standard]==0.135.1
pwdlib[argon2]==0.3.0
pyjwt==2.12.1
sqlmodel==0.0.37